Extract MathJax providers to a new module
